small QOL changes + bugs
1) Disabled right click in the page, so the users don't see the right
click menu.

2)removed the href="#" from the sidebar. When the screen is small and
the href is still there, the page scrolls to the top of the page, we
don't want that.

3) Fixed a bug in assigning distance and size to a finding:

Needed to add an index when getting the findings of one method, to get
the correct finding.

Also needed to add a (click) event to the input box so the
distance/volume also gets updated when using the carets.

Also when typing in input boxes, disabled hotkeys.